pyproject.toml
[tool.poetry]
name = "elevator_storage"
version = "0.1.0"
description = ""
authors = ["guilherme <guilhermerezende770@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.12"
fastapi = "^0.109.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
pytz = "^2023.4"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pymysql = "^1.1.0"
cryptography = "^42.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
